sys/checksum: correct loop variable types
`unsigned` was mixed with `size_t`. On 64 bit platforms, `unsigned`
will can be smaller sized than `size_t`. This can lead to infinite
loops.

`int` was used for simple counter. Prefer `unsigned` here.

Also fix incorrect bracing (Allman style).

+++ sys/checksum/crc32.c
@@ -108,7 +108,7 @@ uint32_t crc32(const void *buf, size_t size)
     const uint8_t *p = buf;
     uint32_t crc = ~0U;
 
-    for (unsigned i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         crc ^= *p++;
         crc = crc32_for_byte(crc);
